requirements.txt
behave==1.2.6
selenium==4.19.0
webdriver-manager==4.0.1
allure-behave==2.13.5
pillow~=10.3.0